requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
firebase-admin==6.2.0
python-dotenv==1.0.0
cryptography==41.0.4
channels==4.0.0
daphne==4.0.0
aiortc==1.5.0
pycryptodome==3.19.0
argon2-cffi==23.1.0
gunicorn==21.2.0 